+++ frontend/src/components/MoTable/MoTableDetail.vue
@@ -2,7 +2,7 @@
   <div>
     <mo-collapse
       class="detail-future"
-      :title="$t('shared.future')"
+      :title="$t('common.future')"
       @show="$emit('show', 'future')">
       <mo-table 
         class="detail-future"
@@ -17,7 +17,7 @@
 
     <mo-collapse
       class="detail-present"
-      :title="$t('shared.present')"
+      :title="$t('common.present')"
       visible
       @show="$emit('show', 'present')">
       <mo-table
@@ -32,7 +32,7 @@
 
     <mo-collapse
       class="detail-past"
-      :title="$t('shared.past')"
+      :title="$t('common.past')"
       @show="$emit('show', 'past')">
       <mo-table 
         :columns="columns"