+++ frontend/src/components/TheLeftMenu.vue
@@ -3,7 +3,7 @@
     <div class="card-body d-flex flex-column">
       <h4 class="card-title">
         <icon name="folder-open"/>
-        {{$t('shared.overview')}}
+        {{$t('common.overview')}}
       </h4>
 
       <div id="tree-wrapper">